+++ source/extensions/isaacsim.robot.manipulators.examples/isaacsim/robot/manipulators/examples/interactive/pick_place/pick_place_example_extension.py
@@ -21,7 +21,7 @@ import omni.ui as ui
 from isaacsim.examples.base.base_sample_extension_experimental import BaseSampleUITemplate
 from isaacsim.examples.browser import get_instance as get_browser_instance
 from isaacsim.gui.components.ui_utils import btn_builder
-from isaacsim.robot.manipulators.examples.franka.interactive.pick_place.pick_place_example import (
+from isaacsim.robot.manipulators.examples.interactive.pick_place.pick_place_example import (
     FrankaPickPlaceInteractive,
 )