+++ source/extensions/isaacsim.examples.interactive/docs/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## [4.13.0] - 2026-01-10
+### Changed 
+- The Getting Started, Replay Follow Target, Robo Factory, and Surface Gripper examples now depend on the new Warp-based APIs
+
 ## [4.12.1] - 2026-01-06
 ### Changed
 - Migrate more events to Events 2.0.