+++ source/extensions/isaacsim.robot.surface_gripper/docs/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## [3.3.7] - 2026-01-10
+### Changed
+- Transfer test cases to core_experimental (Warp-based APIs)
+
 ## [3.3.6] - 2025-12-07
 ### Changed
 - Fix clang tidy issues in cpp code